# Conflicts:
#	Unidad6/Ud06Hoja05Ej01/src/main/java/unidad6/ud06hoja05ej01/Jugador.java

+++ Unidad6/Ud06Hoja03Ej04/src/main/java/unidad6/ud06hoja03ej04/ClientInput.java
@@ -37,7 +37,7 @@ public class ClientInput {
         return clientesNoAtendidos == 0 ? -1 : (double) tiempoEsperaTotal / clientesNoAtendidos;
     }
 
-    public long tiempoEsperaMedio() {
+    public double tiempoEsperaMedio() {
         long tiempoEsperaTotal = 0;
         int clientesAtendidos = 0;
         for (Cliente cliente : clientes) {
